move notification related classes to their own package
