+++ src/main/java/com/iexec/common/notification/TaskNotification.java
@@ -1,4 +1,4 @@
-package com.iexec.common.result;
+package com.iexec.common.notification;
 
 import lombok.*;
